Stop throttling gracefully
Fixes eclipse-vertx/vert.x#4209

When a clustered Vert.x instance is closed, there is a period of time when the Throttling class may still try to submit tasks (because it has pending events) whereas the worker pool has been shutdown, leading to RejectedExecutionException added to the logs.

With this change, we can make sure pending events are all handled before the vert.x worker pool is closed.

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>