Use fixed nodeId from HZ attributes config
Fixes #164

Hazelcast may change the member uuid if it goes out of the grid for a moment (e.g. if suspected to be unhealthy).
We can't use this identifier as Vert.x nodeId.

Instead, we create one and store it as a member attribute.
This has to happen before starting the member because in recent versions of Hazelcast, member attributes can no longer be updated.

As a consequence, users who rely on an existing Hazelcast instance will have to configure the attribute manually (breaking change, documented).

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>